refactor(core): cut the installer scaffolding an audit found
A ponytail pass over the previous commit, applied. Nothing about what the code
does changes; both original bugs are still caught by the same probes.

  * `executable` property -- zero callers anywhere in the tree. Added on the
    theory someone would want it, which is the theory this repo keeps paying
    for.
  * `UvxInstaller` / `NpxInstaller` -- two classes whose whole body bound
    constructor arguments. Now two functions returning a configured
    `CommandRuntimeInstaller`. The `-y` rationale moved with `npx_installer`,
    so the knowledge survives the hierarchy.
  * `installers/` package -- a directory whose `__init__.py` re-exported four
    names from the one module beside it. Now `infrastructure/installers.py`.
    (The emptied directory had to go too: a leftover `__pycache__` makes it a
    namespace package, which would shadow the new module.)
  * `specifier()` -- one caller, its own `install()`, no test. Inlined.
  * the `supports()` guard in `install()`, and the test written for it. The
    handler dispatches through `self._installers[registry_type]`, so a
    mismatched type can only arrive from a caller that does not exist.
    `supports()` itself stays: the port declares it.
  * `runtime_availability`'s nested `ready()` closure -> a set comprehension.

Kept, and worth saying why: the PATH re-check inside `install()` looks like a
duplicate of the bootstrap probe and is not. Bootstrap reads availability once
at start; this one is what makes a week-old gateway say "uvx is not on PATH"
instead of "the subprocess did not start". And `uninstall()` stays a no-op with
no caller, because the port declares it -- removable only with the port.

159 -> 132 lines of source, one file fewer.

Refs #958
